api/requirements.txt
numpy
opencv-python
keras
tensorflow
h5py
scipy
numpy
pillow
python-dotenv
flask